# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DealBattle",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/session": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_v1_session_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_v1_session_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Expire Session",
        "operationId": "expire_session_api_v1_session_delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "summary": "Products",
        "operationId": "products_api_v1_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductView"
                  },
                  "type": "array",
                  "title": "Response Products Api V1 Products Get"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/v1/deals": {
      "post": {
        "summary": "Create Deal",
        "operationId": "create_deal_api_v1_deals_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeal"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealView"
                }
              }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED / CREDENTIAL_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "403": {
            "description": "CSRF_INVALID / ORIGIN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "404": {
            "description": "DEAL_NOT_FOUND / PRODUCT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "409": {
            "description": "VERSION_CONFLICT / STATE_CONFLICT / APPROVAL_STALE / IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          },
          "503": {
            "description": "DATABASE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorView": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorView"
      },
...
      "ProductView": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "product_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Product Version"
          },
          "seller_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Seller Version"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "ram_gb": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Ram Gb"
          },
          "ssd_gb": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Ssd Gb"
          },
          "asking_price_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Asking Price Krw"
          },
          "shipping_fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Shipping Fee Krw"
          },
          "fee_krw": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Fee Krw"
          },
          "delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Delivery Date"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "source": {
            "type": "string",
            "const": "simulated",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "seller_id",
          "product_version",
          "seller_version",
          "name",
          "ram_gb",
          "ssd_gb",
          "asking_price_krw",
          "shipping_fee_krw",
          "fee_krw",
          "delivery_date",
          "stock",
          "expires_at",
          "source"
        ],
        "title": "ProductView"
      },
...
      "SessionInput": {
        "properties": {
          "credential": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionInput"
      },
      "SessionView": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "csrf_token",
          "expires_at",
          "mode"
        ],
        "title": "SessionView"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dealbattle_session"
      }
    }
  }
}